refactor : PurchaseResponse -> GalaxiaPurchaseResponse 클래스 이름 수정 및 외부 시스템 클라이언트 디렉토리로 이동
