+++ src/main/java/com/pintoss/auth/common/client/billgate/PurchaseApiClient.java
@@ -6,7 +6,6 @@ import com.galaxia.api.crypto.GalaxiaCipher;
 import com.galaxia.api.crypto.Seed;
 import com.galaxia.api.util.NumberUtil;
 import com.pintoss.auth.module.order.integration.PurchaseRequestBuilder;
-import com.pintoss.auth.module.order.integration.PurchaseResponse;
 import com.pintoss.auth.module.payment.application.PaymentMethodType;
 import java.util.Arrays;
 import lombok.extern.slf4j.Slf4j;
@@ -27,7 +26,7 @@ public class PurchaseApiClient {
         this.IV = properties.getSecret().getIv();
     }
 
-    public PurchaseResponse purchase(String orderNo, String transactionId, String mid, Long amount, PaymentMethodType paymentMethodType, Long salePrice, String productCode) {
+    public GalaxiaPurchaseResponse purchase(String orderNo, String transactionId, String mid, Long amount, PaymentMethodType paymentMethodType, Long salePrice, String productCode) {
         try {
             String requestHeader = PurchaseRequestBuilder.buildHeader(orderNo);
             String bodyPlain = PurchaseRequestBuilder.buildBody(orderNo,transactionId, mid,salePrice.toString(), paymentMethodType, salePrice.toString(), productCode);
@@ -61,7 +60,7 @@ public class PurchaseApiClient {
             byte[] cleanBytes = Arrays.copyOfRange(decryptedBytes, 0, length);
             String plainBody = new String(cleanBytes, "EUC-KR");
             log.info("[DEBUG] 복호화 결과 (plainBody): " + plainBody);
-            return PurchaseResponse.fromBytes(cleanBytes);
+            return GalaxiaPurchaseResponse.fromBytes(cleanBytes);
         } catch (Exception e) {
             e.printStackTrace();
             return null;
