+++ src/main/java/com/pintoss/auth/module/order/integration/VoucherPurchaseService.java
@@ -1,7 +1,8 @@
 package com.pintoss.auth.module.order.integration;
 
+import com.pintoss.auth.common.client.billgate.GalaxiaPurchaseResponse;
 import com.pintoss.auth.module.payment.application.PaymentMethodType;
 
 public interface VoucherPurchaseService {
-    PurchaseResponse purchase(String orderId, String transactionId, String mid, Long taxAmount, PaymentMethodType paymentMethodType);
+    GalaxiaPurchaseResponse purchase(String orderId, String transactionId, String mid, Long taxAmount, PaymentMethodType paymentMethodType);
 }
